+++ NotifyMeStart/src/main/java/be/xplore/notifyme/StartNotifyMe.java
@@ -2,7 +2,6 @@ package be.xplore.notifyme;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.PropertySource;
 
 /**
