+++ NotifyMeStart/src/main/resources/application.properties
@@ -19,3 +19,5 @@ spring.jpa.database-platform=org.hibernate.dialect.PostgreSQLDialect
 spring.datasource.url=jdbc:postgresql://localhost:5432/notifyme
 spring.datasource.username=admin
 spring.jpa.show-sql=true
+
+twilio.account.phoneno=+12062071090
