+++ domain/src/main/java/be/xplore/notifyme/persistence/ICommunicationPreferenceRepo.java
@@ -1,7 +1,7 @@
 package be.xplore.notifyme.persistence;
 
 import be.xplore.notifyme.domain.CommunicationPreference;
-import be.xplore.notifyme.domain.communicationstrategies.ICommunicationStrategy;
+import be.xplore.notifyme.services.communicationstrategies.ICommunicationStrategy;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.stereotype.Repository;
