+++ domain/src/main/java/be/xplore/notifyme/services/CommunicationPreferenceService.java
@@ -1,9 +1,9 @@
 package be.xplore.notifyme.services;
 
 import be.xplore.notifyme.domain.CommunicationPreference;
-import be.xplore.notifyme.domain.communicationstrategies.EmailCommunicationStrategy;
-import be.xplore.notifyme.domain.communicationstrategies.ICommunicationStrategy;
-import be.xplore.notifyme.domain.communicationstrategies.SmsCommunicationStrategy;
+import be.xplore.notifyme.services.communicationstrategies.EmailCommunicationStrategy;
+import be.xplore.notifyme.services.communicationstrategies.ICommunicationStrategy;
+import be.xplore.notifyme.services.communicationstrategies.SmsCommunicationStrategy;
 import be.xplore.notifyme.persistence.ICommunicationPreferenceRepo;
 import java.util.List;
 import javax.validation.ValidationException;
