+++ domain/src/main/java/be/xplore/notifyme/services/conversion/CommunicationStrategyConverter.java
@@ -1,10 +1,10 @@
 package be.xplore.notifyme.services.conversion;
 
-import be.xplore.notifyme.domain.communicationstrategies.EmailCommunicationStrategy;
-import be.xplore.notifyme.domain.communicationstrategies.ICommunicationStrategy;
-import be.xplore.notifyme.domain.communicationstrategies.IEmailService;
-import be.xplore.notifyme.domain.communicationstrategies.ISmsService;
-import be.xplore.notifyme.domain.communicationstrategies.SmsCommunicationStrategy;
+import be.xplore.notifyme.services.communicationstrategies.EmailCommunicationStrategy;
+import be.xplore.notifyme.services.communicationstrategies.ICommunicationStrategy;
+import be.xplore.notifyme.services.communicationstrategies.IEmailService;
+import be.xplore.notifyme.services.communicationstrategies.ISmsService;
+import be.xplore.notifyme.services.communicationstrategies.SmsCommunicationStrategy;
 import javax.persistence.AttributeConverter;
 import lombok.RequiredArgsConstructor;
