+++ domain/src/test/java/be/xplore/notifyme/services/NotificationServiceTest.java
@@ -12,9 +12,9 @@ import be.xplore.notifyme.domain.CommunicationPreference;
 import be.xplore.notifyme.domain.Message;
 import be.xplore.notifyme.domain.Notification;
 import be.xplore.notifyme.domain.User;
-import be.xplore.notifyme.domain.communicationstrategies.EmailCommunicationStrategy;
-import be.xplore.notifyme.domain.communicationstrategies.IEmailService;
-import be.xplore.notifyme.domain.communicationstrategies.ISmsService;
+import be.xplore.notifyme.services.communicationstrategies.EmailCommunicationStrategy;
+import be.xplore.notifyme.services.communicationstrategies.IEmailService;
+import be.xplore.notifyme.services.communicationstrategies.ISmsService;
 import be.xplore.notifyme.persistence.IMessageRepo;
 import be.xplore.notifyme.persistence.INotificationRepo;
 import java.util.ArrayList;
