+++ ses/src/main/java/be/xplore/notifyme/services/SesService.java
@@ -1,7 +1,7 @@
 package be.xplore.notifyme.services;
 
 import be.xplore.notifyme.config.SesConfig;
-import be.xplore.notifyme.domain.communicationstrategies.IEmailService;
+import be.xplore.notifyme.services.communicationstrategies.IEmailService;
 import java.util.Properties;
 import javax.mail.Message;
 import javax.mail.MessagingException;
