--- sms-service/src/main/resources/application.properties
@@ -1,1 +1,0 @@
-twilio.phone.no=+12062071090
